Complete base build
Add clean up line to file; no need to copy the "shiny-server" file because this isn't necessary (from Rafal's Dockerfile). 

Not sure what the main reasons behind using phusion are; I don't really see the need for it (SSH is cool, but we will never need it in practice - the iNZight developers won't have access anyway).